Added using items from EQ, added class Item_medicine
You use item by clicking on it. When used it disappears from Equipement.
For some reason sometimes when you click on item it doesn't work...
Probably has something to do with optimalisation of Events. 

class Item_medicine can be used for creating hp altering items. As of now there is painkiller and small poison.